requirements.txt
asgiref==3.2.7
Django==3.0.6
django-cors-headers==3.3.0
django-filter==2.2.0
djangorestframework==3.11.0
mysqlclient==1.4.6
Pygments==2.6.1
pytz==2020.1
sqlparse==0.3.1
